+++ src/giftCards/GiftCardsList/GiftCardListPage.tsx
@@ -1,15 +1,17 @@
+import VerticalSpacer from "@saleor/apps/components/VerticalSpacer";
 import Container from "@saleor/components/Container";
 import React from "react";
 
 import GiftCardsListHeader from "./GiftCardsListHeader";
-import GiftCardsListOrderCard from "./GiftCardsListOrderCard/GiftCardsListOrderCard";
+import GiftCardsListOrderInfoCard from "./GiftCardsListOrderInfoCard/GiftCardsListOrderInfoCard";
 import GiftCardsListTable from "./GiftCardsListTable";
 
 const GiftCardsListPage: React.FC = () => (
   <Container>
     <GiftCardsListHeader />
     <GiftCardsListTable />
-    <GiftCardsListOrderCard />
+    <VerticalSpacer spacing={2} />
+    <GiftCardsListOrderInfoCard />
   </Container>
 );
