--- src/giftCards/GiftCardsList/GiftCardsListOrderCard/messages.ts
@@ -1,9 +1,0 @@
-import { defineMessages } from "react-intl";
-
-export const giftCardListOrderCardMessages = defineMessages({
-  giftCardOrderMessage: {
-    defaultMessage:
-      "Gift cards will appear after their order is fullfilled. {orderPageLink}",
-    description: "alert card message"
-  }
-});
