+++ src/giftCards/GiftCardsList/GiftCardsListOrderInfoCard/messages.ts
@@ -1,0 +1,12 @@
+import { defineMessages } from "react-intl";
+
+export const giftCardListOrderCardMessages = defineMessages({
+  giftCardOrderInfoMessage: {
+    defaultMessage: "Gift cards will appear after their order is fullfilled. ",
+    description: "alert card message"
+  },
+  giftCardOrderInfoMessageLink: {
+    defaultMessage: "View Orders with Gift Cards",
+    description: "alert card message link"
+  }
+});
